requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.7
numpy==1.26.3
pytest==7.4.4